feat: update Quarkus to 3.31.2

+++ stackgres-k8s/src/restapi/src/main/java/io/stackgres/apiweb/dto/dbops/DbOpsPgbenchStatusTransactionsPerSecondOverTime.java
@@ -20,7 +20,7 @@ public class DbOpsPgbenchStatusTransactionsPerSecondOverTime {
 
   private String valuesUnit;
 
-  private Integer intervalDuration;
+  private BigDecimal intervalDuration;
 
   private String intervalDurationUnit;
 
@@ -40,11 +40,11 @@ public class DbOpsPgbenchStatusTransactionsPerSecondOverTime {
     this.valuesUnit = valuesUnit;
   }
 
-  public Integer getIntervalDuration() {
+  public BigDecimal getIntervalDuration() {
     return intervalDuration;
   }
 
-  public void setIntervalDuration(Integer intervalDuration) {
+  public void setIntervalDuration(BigDecimal intervalDuration) {
     this.intervalDuration = intervalDuration;
   }
